fix memory issue

+++ df_psm/psm.h
@@ -7,6 +7,6 @@
 #include "src/file.h"
 
 typedef uint8_t byte;
-typedef uint16_t word;
+typedef uint32_t word;
 
 #endif
